+++ src/test/e2e/services/reset-token-service.e2e.test.ts
@@ -1,4 +1,3 @@
-import test from 'ava';
 import dbInit from '../helpers/database-init';
 import getLogger from '../../fixtures/no-logger';
 import ResetTokenService from '../../../lib/services/reset-token-service';
@@ -10,6 +9,7 @@ import User from '../../../lib/types/user';
 import { IUnleashConfig } from '../../../lib/types/option';
 import { createTestConfig } from '../../config/test-config';
 import SessionService from '../../../lib/services/session-service';
+import InvalidTokenError from '../../../lib/error/invalid-token-error';
 
 const config: IUnleashConfig = createTestConfig();
 
@@ -22,7 +22,7 @@ let accessService: AccessService;
 let userService: UserService;
 let resetTokenService: ResetTokenService;
 let sessionService: SessionService;
-test.before(async () => {
+beforeAll(async () => {
     db = await dbInit('reset_token_service_serial', getLogger);
     stores = db.stores;
     accessService = new AccessService(stores, config);
@@ -49,69 +49,65 @@ test.before(async () => {
     userIdToCreateResetFor = userToCreateResetFor.id;
 });
 
-test.after.always(async () => {
-    db.destroy();
+afterAll(async () => {
+    if (db) {
+        await db.destroy();
+    }
 });
 
-test.serial('Should create a reset link', async t => {
+test('Should create a reset link', async () => {
     const url = await resetTokenService.createResetPasswordUrl(
         userIdToCreateResetFor,
         adminUser,
     );
 
-    t.is(
-        url.toString().substring(0, url.toString().indexOf('=')),
+    expect(url.toString().substring(0, url.toString().indexOf('='))).toBe(
         `${config.server.unleashUrl}/reset-password?token`,
     );
 });
 
-test.serial(
-    'Should create a reset link with unleashUrl with context path',
-    async t => {
-        const localConfig = createTestConfig({
-            server: { unleashUrl: 'http://localhost:4242/my/sub/path' },
-        });
-        const resetToken: ResetTokenService = new ResetTokenService(
-            stores,
-            localConfig,
-        );
+test('Should create a reset link with unleashUrl with context path', async () => {
+    const localConfig = createTestConfig({
+        server: { unleashUrl: 'http://localhost:4242/my/sub/path' },
+    });
+    const resetToken: ResetTokenService = new ResetTokenService(
+        stores,
+        localConfig,
+    );
 
-        const url = await resetToken.createResetPasswordUrl(
-            userIdToCreateResetFor,
-            adminUser,
-        );
-        t.is(
-            url.toString().substring(0, url.toString().indexOf('=')),
-            `${localConfig.server.unleashUrl}/reset-password?token`,
-        );
-    },
-);
+    const url = await resetToken.createResetPasswordUrl(
+        userIdToCreateResetFor,
+        adminUser,
+    );
+    expect(url.toString().substring(0, url.toString().indexOf('='))).toBe(
+        `${localConfig.server.unleashUrl}/reset-password?token`,
+    );
+});
 
-test.serial('Should create a welcome link', async t => {
+test('Should create a welcome link', async () => {
     const url = await resetTokenService.createNewUserUrl(
         userIdToCreateResetFor,
         adminUser.username,
     );
     const urlS = url.toString();
-    t.is(
-        urlS.substring(0, urlS.indexOf('=')),
+    expect(urlS.substring(0, urlS.indexOf('='))).toBe(
         `${config.server.unleashUrl}/new-user?token`,
     );
 });
 
-test.serial('Tokens should be one-time only', async t => {
+test('Tokens should be one-time only', async () => {
     const token = await resetTokenService.createToken(
         userIdToCreateResetFor,
         adminUser,
     );
 
     const accessGranted = await resetTokenService.useAccessToken(token);
-    t.is(accessGranted, true);
+    expect(accessGranted).toBe(true);
     const secondGo = await resetTokenService.useAccessToken(token);
-    t.is(secondGo, false);
+    expect(secondGo).toBe(false);
 });
 
-test.serial('Creating a new token should expire older tokens', async t => {
+test('Creating a new token should expire older tokens', async () => {
     const firstToken = await resetTokenService.createToken(
         userIdToCreateResetFor,
         adminUser,
@@ -120,24 +116,23 @@ test.serial('Creating a new token should expire older tokens', async t => {
         userIdToCreateResetFor,
         adminUser,
     );
-    await t.throwsAsync<NotFoundError>(async () =>
+    await expect(async () =>
         resetTokenService.isValid(firstToken.token),
-    );
+    ).rejects.toThrow(InvalidTokenError);
     const validToken = await resetTokenService.isValid(secondToken.token);
-    t.is(secondToken.token, validToken.token);
+    expect(secondToken.token).toBe(validToken.token);
 });
 
-test.serial(
-    'Retrieving valid invitation links should retrieve an object with userid key and token value',
-    async t => {
-        const token = await resetTokenService.createToken(
-            userIdToCreateResetFor,
-            adminUser,
-        );
-        t.truthy(token);
-        const activeInvitations = await resetTokenService.getActiveInvitations();
-        t.true(Object.keys(activeInvitations).length === 1);
-        t.true(+Object.keys(activeInvitations)[0] === userIdToCreateResetFor);
-        t.truthy(activeInvitations[userIdToCreateResetFor]);
-    },
-);
+test('Retrieving valid invitation links should retrieve an object with userid key and token value', async () => {
+    const token = await resetTokenService.createToken(
+        userIdToCreateResetFor,
+        adminUser,
+    );
+    expect(token).toBeTruthy();
+    const activeInvitations = await resetTokenService.getActiveInvitations();
+    expect(Object.keys(activeInvitations).length === 1).toBe(true);
+    expect(+Object.keys(activeInvitations)[0] === userIdToCreateResetFor).toBe(
+        true,
+    );
+    expect(activeInvitations[userIdToCreateResetFor]).toBeTruthy();
+});
